Move rune locations, fix Honor Among Thieves input
