+++ ui/rogue/inputs.ts
@@ -1,8 +1,7 @@
-import { RogueRune } from 'ui/core/proto/rogue.js';
-
 import * as InputHelpers from '../core/components/input_helpers.js';
 import { Player } from '../core/player.js';
 import { ItemSlot, Spec } from '../core/proto/common.js';
+import { RogueRune } from '../core/proto/rogue.js';
 
 
 // Configuration for spec-specific UI elements on the settings tab.
@@ -12,5 +11,4 @@ export const HonorOfThievesCritRate = InputHelpers.makeSpecOptionsNumberInput<Sp
 	fieldName: 'honorAmongThievesCritRate',
 	label: 'Honor Among Thieves Crit Rate',
 	labelTooltip: 'Number of crits other group members generate within 100 seconds',
-	showWhen: (player: Player<Spec.SpecRogue>) => player.getEquippedItem(ItemSlot.ItemSlotHead)?.rune?.id == RogueRune.RuneHonorAmongThieves,
 });
